requirements.txt
Django>=4.0
python-dotenv
spotipy
lyricsgenius